requirements.txt
flask>=3.0.3
flask-cors>=5.0.0
tensorflow>=2.16.1
numpy>=1.26.4
Pillow>=10.4.0
scikit-learn>=1.5.1
pandas>=2.2.2
joblib>=1.4.2
